requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
deprecation==2.1.0
distro==1.9.0
frozenlist==1.5.0
geojson==2.5.0
gotrue==2.11.4
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
jiter==0.8.2
multidict==6.1.0
numpy==2.2.3
openai==1.64.0
packaging==24.2
pandas==2.2.3
postgrest==0.19.3
propcache==0.3.0
pydantic==2.10.6
pydantic_core==2.27.2
pyown==0.3.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
realtime==2.4.0
requests==2.32.3
six==1.17.0
sniffio==1.3.1
storage3==0.11.3
StrEnum==0.4.15
supafunc==0.9.3
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
websockets==14.2
yarl==1.18.3
